+++ vmware_horizon.py
@@ -472,17 +472,116 @@ class Inventory:
                 return response.json()
 
     def disconnect_sessions(self, session_ids: list):
-        """Disconnects user sessions
+        """Disconnects user sessions.
 
         Requires list of session ids
         Available for Horizon 8 2006 and later."""
         headers = self.access_token
         headers["Content-Type"] = 'application/json'
         json_data = json.dumps(session_ids)
-        print(json_data)
-        response = requests.post(f'{self.url}/rest​/inventory​/v1​/sessions​/action​/disconnect', verify=False,  headers=headers, data=json_data)
+        response = requests.post(f'{self.url}/rest/inventory/v1/sessions/action/disconnect', verify=False,  headers=headers, data=json_data)
         if response.status_code == 400:
             error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 404:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        elif response.status_code != 200:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.json()
+
+    def logoff_sessions(self, session_ids: list, forced_logoff:bool=False):
+        """Logs user sessions off.
+
+        Requires list of session ids optional to set forced to True to force a log off (defaults to False)
+        Available for Horizon 8 2006 and later."""
+        headers = self.access_token
+        headers["Content-Type"] = 'application/json'
+        json_data = json.dumps(session_ids)
+        response = requests.post(f'{self.url}/rest/inventory/v1/sessions/action/logoff?forced={forced_logoff}', verify=False,  headers=headers, data=json_data)
+        if response.status_code == 400:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 404:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        elif response.status_code != 200:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.json()
+
+    def reset_session_machines(self, session_ids: list):
+        """Resets machine of user sessions. The machine must be managed by Virtual Center and the session cannot be an application or an RDS desktop session.
+
+        Requires list of session ids
+        Available for Horizon 8 2006 and later."""
+        headers = self.access_token
+        headers["Content-Type"] = 'application/json'
+        json_data = json.dumps(session_ids)
+        response = requests.post(f'{self.url}/rest/inventory/v1/sessions/action/reset', verify=False,  headers=headers, data=json_data)
+        if response.status_code == 400:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 404:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        elif response.status_code != 200:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.json()
+
+    def restart_session_machines(self, session_ids: list):
+        """Restarts machine of user sessions. The machine must be managed by Virtual Center and the session cannot be an application or an RDS desktop session.
+
+        Requires list of session ids
+        Available for Horizon 8 2006 and later."""
+        headers = self.access_token
+        headers["Content-Type"] = 'application/json'
+        json_data = json.dumps(session_ids)
+        response = requests.post(f'{self.url}/rest/inventory/v1/sessions/action/reset', verify=False,  headers=headers, data=json_data)
+        if response.status_code == 400:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 404:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        elif response.status_code != 200:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.json()
+
+    def send_message_sessions(self, session_ids: list, message:str, message_type:str="INFO"):
+        """Sends the message to user sessions
+
+        Requires list of session ids, message type (INFO,WARNING,ERROR) and a message
+        Available for Horizon 8 2006 and later."""
+        headers = self.access_token
+        headers["Content-Type"] = 'application/json'
+        data={}
+        data["message"] = message
+        data["message_type"] = message_type
+        data["session_ids"] = session_ids
+        json_data = json.dumps(data)
+        response = requests.post(f'{self.url}/rest/inventory/v1/sessions/action/send-message', verify=False,  headers=headers, data=json_data)
+        if response.status_code == 400:
+            error_message = (response.json())["error_messages"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
             raise Exception(f"Error {response.status_code}: {response.reason}")